Fix compilation

+++ src/open_lambda.rs
@@ -1,3 +1,5 @@
+use crate::Error;
+
 mod api {
     #[link(wasm_import_module = "ol_args")]
     extern "C" {
@@ -7,7 +9,7 @@ mod api {
 
 pub(crate) fn getrandom_inner(dest: &mut [u8]) -> Result<(), Error> {
     let ptr = dest.as_mut_ptr();
-    let len = dest.len();
+    let len = dest.len() as u32;
 
     unsafe {
         api::get_random_value(dest, len);
